Box tree refactor
